change some names, seperate generic game mechanics, ordered components and systems (slightly) more logically
